feat: removed editor logic and refactored
