+++ Editor/UI/EditorWindows/Templates/AppIdTemplate.cs
@@ -60,7 +60,7 @@ namespace ReadyPlayerMe.Core.Editor
                 AnalyticsEditorLogger.EventLogger.LogUpdatePartnerURL(id, appIdField.value);
             }
 
-            CoreSettingsHandler.SaveAppId(appIdField.value);
+            CoreSettingsManager.SaveAppId(appIdField.value);
         }
     }
 }
